Ground Detection Clean
Cleaned the ground detection script PES_Grounded.cs, first so that it only deals with ground detection, and no longer implements the step climber functionality.
We will instead make the step climber an independent script which will disable the whole PES_Grounded.cs script when enabled, and set the grounded state to true.

The ground Collider storage and update events have also been removed, since in this new version of the project, we won't be using player material friction, and these were used to deal with the walls friction adaptation.