PIS Combat Refactor
PIS Combat no longer exists, and is now devided in two separated scripts, PI_AMapsManager.cs and PIA_RunningProcessing.cs

PIS Combat had a weird behavior, and multiple purposes.
Trying to make sure that the playerInputActions exists with events that we would subscribe to PIS Combat so that we can then subscribe to the playerInputActions maps' actions was pointless.

--------
We know (in this version at least) that the playerInputActions will always exists.
It is now instantiated in the PI_AMapsManager Awake method, a script that will always be in the scene. It allows us to always be able to access it in a non null state in the start methods of our actions scripts.
This newly created script will deal with action maps through the game in general.
Right now we only focus on Arena state, so it only initate the playerInputActions and enables the Arena Action Map on Enable.

---------
The PIS combat would also deal with the basic WASD inputs. I switched this functionality in its own script, PIA_RunningProcessing.

PIS combat no longer exists.