Walk bobbing + Mid Air sprinting Animation
bobbing now includes a walking setting.
It depends on an event casted by PMA_GroundControlManager when reaching a certain speed on ground.

When the player leaves the ground while sprinting, it plays a mid air animation.
It depends on an event casted by PMA_GroundControlManager when leaving the ground while sprinting.