Mid Air Stop Sprint input conflict Fix
StopAirSprint wouldn't be cast when landing and not sprinting anymore, if the player entered sprint canceling input in air.
Now we cast it when landing, no matter if the player is sprinting or not.
I could also cast the event as a response to the StopOrLess event, but I prefer keeping the animation on until grounded.