Speed Accel and Drag Modifiers & Jump Modifier
Implemented a way to add independant modifiers to the applied speed accel and drag through dictionnaries in PM_SC_Manager.
PMA_Jump introduces a speed penalty when space bar is held down.
We might want to apply this penalty after a short amount of time instead of systematically applying it when pressing spacebar, so the penalty is never applied when you briefly tap space bar.