Air Drifting
Added an air drift feature in PMA_Lurch. Its behavior is similar to tap strafing, but it doesn't require to smash forward.
Instead, we add a speed and accel modifier when the player presses forward mid air for the remaining lurch time, and changes the way PM_SC_Manager applies the velocity changes by clamping it so that you cannot gain speed out of this drifting mode.