Jump Input Buffer
Now, PIA_JumpProcessing deals with jump inputs, and creates an external Buffer.
Previously, a buffer was implemented directly in the PMA_Jump script. It was clunky, and it meant it was specific for the basic jump. Making it an external feature allows it to be used in the future wall interactions.

To Check for a buffered input, modules will have to use PIA_JumpProcessing.UseBuffer() as a boolean test to apply the corresponding action.

For example, PMA_Jump just responds directly to normal jump inputs, but also call UseBuffer() on landing.

TO BE TESTED :
Be careful as it might have introduced a bug where it is possible to trigger two stacked jump forces if you jump on the exact tick you're landing.