Jump Stun skip fix + Slide Stun skip fix
Now, no jump is shot if the player was holding space bar. However, he can skip the stun by tapping jump right before landing, in the buffer time, or jumping slightly after, as soon as possible to reduce the friction impact.
If the player is sliding, the drag penalty is greater.